+++ CHANGELOG.md
@@ -6,4 +6,4 @@
 - Retrieve data from 3rd party API by describing URL, headers, query parameters and authentication with templates
 - Custom Dataset for simple data retrievals
 - Custom Recipe using an input dataset as variables for your templates
-- Implements basic authentication, bearer token, API key, OAuth 2.0 token
+- Implements basic authentication, bearer token, API key